【I4HG7L】【AppStore】App变现后台实现 --- 增加虚机部署参数

+++ src/main/java/org/edgegallery/appstore/domain/model/order/OrderRepository.java
@@ -28,7 +28,7 @@ public interface OrderRepository {
 
     List<Order> queryOrders(Map<String, Object> params);
 
-    long getCountByCondition(Map<String, Object> params);
+    Integer getCountByCondition(Map<String, Object> params);
 
     String maxOrderNum();
